Split events.rs into multiple files and start work on FilterEvents
