+++ subxt/src/events/events.rs
@@ -14,50 +14,29 @@
 // You should have received a copy of the GNU General Public License
 // along with subxt.  If not, see <http://www.gnu.org/licenses/>.
 
-//! For working with events.
+//! A representation of a block of events.
 
 use crate::{
     error::BasicError,
-    metadata::MetadataError,
     Client,
     Config,
     Event,
     Metadata,
     Phase,
 };
-use bitvec::{
-    order::Lsb0,
-    vec::BitVec,
-};
 use codec::{
-    Codec,
     Compact,
     Decode,
     Error as CodecError,
     Input,
 };
 use derivative::Derivative;
-use futures::{
-    Future,
-    FutureExt,
-    Stream,
-    StreamExt,
-};
-use jsonrpsee::core::client::Subscription;
-use scale_info::{
-    PortableRegistry,
-    TypeDef,
-    TypeDefPrimitive,
-};
 use sp_core::{
-    storage::StorageKey,
-    twox_128,
     Bytes,
+    twox_128,
+    storage::StorageKey,
 };
-use std::{
-    marker::Unpin,
-    task::Poll,
-};
+use super::decoding;
 
 /// Obtain events at some block hash. The generic parameter is what we
 /// will attempt to decode each event into if using [`Events::iter()`],
@@ -100,136 +79,18 @@ pub async fn at<T: Config, Evs: Decode>(
     })
 }
 
-/// Subscribe to events from blocks.
-///
-/// **Note:** these blocks haven't necessarily been finalised yet; prefer
-/// [`Events::subscribe_finalized()`] if that is important.
-///
-/// **Note:** This function is hidden from the documentation
-/// and is exposed only to be called via the codegen. Thus, prefer to use
-/// `api.events().subscribe()` over calling this directly.
-#[doc(hidden)]
-pub async fn subscribe<T: Config, Evs: Decode + 'static>(
-    client: &'_ Client<T>,
-) -> Result<EventSubscription<'_, T, Evs>, BasicError> {
-    let block_subscription = client.rpc().subscribe_blocks().await?;
-    Ok(EventSubscription::new(client, block_subscription))
-}
-
-/// Subscribe to events from finalized blocks.
-///
-/// **Note:** This function is hidden from the documentation
-/// and is exposed only to be called via the codegen. Thus, prefer to use
-/// `api.events().subscribe_finalized()` over calling this directly.
-#[doc(hidden)]
-pub async fn subscribe_finalized<T: Config, Evs: Decode + 'static>(
-    client: &'_ Client<T>,
-) -> Result<EventSubscription<'_, T, Evs>, BasicError> {
-    let block_subscription = client.rpc().subscribe_finalized_blocks().await?;
-    Ok(EventSubscription::new(client, block_subscription))
-}
-
-/// A subscription to events that implements [`Stream`], and returns [`Events`] objects for each block.
-#[derive(Derivative)]
-#[derivative(Debug(bound = ""))]
-pub struct EventSubscription<'a, T: Config, Evs: Decode + 'static> {
-    finished: bool,
-    client: &'a Client<T>,
-    block_header_subscription: Subscription<T::Header>,
-    #[derivative(Debug = "ignore")]
-    at: Option<
-        std::pin::Pin<
-            Box<dyn Future<Output = Result<Events<'a, T, Evs>, BasicError>> + 'a>,
-        >,
-    >,
-    _event_type: std::marker::PhantomData<Evs>,
-}
-
-impl<'a, T: Config, Evs: Decode> EventSubscription<'a, T, Evs> {
-    fn new(
-        client: &'a Client<T>,
-        block_header_subscription: Subscription<T::Header>,
-    ) -> Self {
-        EventSubscription {
-            finished: false,
-            client,
-            block_header_subscription,
-            at: None,
-            _event_type: std::marker::PhantomData,
-        }
-    }
-}
-
-impl<'a, T: Config, Evs: Decode> Unpin for EventSubscription<'a, T, Evs> {}
-
-// We want `EventSubscription` to implement Stream. The below implementation is the rather verbose
-// way to roughly implement the following function:
-//
-// ```
-// fn subscribe_events<T: Config, Evs: Decode>(client: &'_ Client<T>, block_sub: Subscription<T::Header>) -> impl Stream<Item=Result<Events<'_, T, Evs>, BasicError>> + '_ {
-//     use futures::StreamExt;
-//     block_sub.then(move |block_header_res| async move {
-//         use sp_runtime::traits::Header;
-//         let block_header = block_header_res?;
-//         let block_hash = block_header.hash();
-//         at(client, block_hash).await
-//     })
-// }
-// ```
-//
-// The advantage of this manual implementation is that we have a named type that we (and others)
-// can derive things on, store away, alias etc.
-impl<'a, T: Config, Evs: Decode> Stream for EventSubscription<'a, T, Evs> {
-    type Item = Result<Events<'a, T, Evs>, BasicError>;
-
-    fn poll_next(
-        mut self: std::pin::Pin<&mut Self>,
-        cx: &mut std::task::Context<'_>,
-    ) -> std::task::Poll<Option<Self::Item>> {
-        // We are finished; return None.
-        if self.finished {
-            return Poll::Ready(None)
-        }
-
-        // If there isn't an `at` function yet that's busy resolving a block hash into
-        // some event details, then poll the block header subscription to get one.
-        if self.at.is_none() {
-            match futures::ready!(self.block_header_subscription.poll_next_unpin(cx)) {
-                None => {
-                    self.finished = true;
-                    return Poll::Ready(None)
-                }
-                Some(Err(e)) => {
-                    self.finished = true;
-                    return Poll::Ready(Some(Err(e.into())))
-                }
-                Some(Ok(block_header)) => {
-                    use sp_runtime::traits::Header;
-                    // Note [jsdw]: We may be able to get rid of the per-item allocation
-                    // with https://github.com/oblique/reusable-box-future.
-                    self.at = Some(Box::pin(at(self.client, block_header.hash())));
-                    // Continue, so that we poll this function future we've just created.
-                }
-            }
-        }
-
-        // If we get here, there will be an `at` function stored. Unwrap it and poll it to
-        // completion to get our events, throwing it away as soon as it is ready.
-        let at_fn = self
-            .at
-            .as_mut()
-            .expect("'at' function should have been set above'");
-        let events = futures::ready!(at_fn.poll_unpin(cx));
-        self.at = None;
-        Poll::Ready(Some(events))
-    }
+// The storage key needed to access events.
+fn system_events_key() -> StorageKey {
+    let mut storage_key = twox_128(b"System").to_vec();
+    storage_key.extend(twox_128(b"Events").to_vec());
+    StorageKey(storage_key)
 }
 
 /// A collection of events obtained from a block, bundled with the necessary
 /// information needed to decode and iterate over them.
 #[derive(Derivative)]
 #[derivative(Debug(bound = ""))]
-pub struct Events<'a, T: Config, Evs: Decode> {
+pub struct Events<'a, T: Config, Evs> {
     metadata: &'a Metadata,
     block_hash: T::Hash,
     // Note; raw event bytes are prefixed with a Compact<u32> containing
@@ -457,7 +318,7 @@ fn decode_raw_event_details<T: Config>(
         let type_id = arg.ty().id();
         let all_bytes = *input;
         // consume some bytes, moving the cursor forward:
-        decode_and_consume_type(type_id, &metadata.runtime_metadata().types, input)?;
+        decoding::decode_and_consume_type(type_id, &metadata.runtime_metadata().types, input)?;
         // count how many bytes were consumed based on remaining length:
         let consumed_len = all_bytes.len() - input.len();
         // move those consumed bytes to the output vec unaltered:
@@ -480,222 +341,14 @@ fn decode_raw_event_details<T: Config>(
     })
 }
 
-// The storage key needed to access events.
-fn system_events_key() -> StorageKey {
-    let mut storage_key = twox_128(b"System").to_vec();
-    storage_key.extend(twox_128(b"Events").to_vec());
-    StorageKey(storage_key)
-}
-
-// Given a type Id and a type registry, attempt to consume the bytes
-// corresponding to that type from our input.
-fn decode_and_consume_type(
-    type_id: u32,
-    types: &PortableRegistry,
-    input: &mut &[u8],
-) -> Result<(), BasicError> {
-    let ty = types
-        .resolve(type_id)
-        .ok_or(MetadataError::TypeNotFound(type_id))?;
-
-    fn consume_type<T: Codec>(input: &mut &[u8]) -> Result<(), BasicError> {
-        T::decode(input)?;
-        Ok(())
-    }
-
-    match ty.type_def() {
-        TypeDef::Composite(composite) => {
-            for field in composite.fields() {
-                decode_and_consume_type(field.ty().id(), types, input)?
-            }
-            Ok(())
-        }
-        TypeDef::Variant(variant) => {
-            let variant_index = u8::decode(input)?;
-            let variant = variant
-                .variants()
-                .iter()
-                .find(|v| v.index() == variant_index)
-                .ok_or_else(|| {
-                    BasicError::Other(format!("Variant {} not found", variant_index))
-                })?;
-            for field in variant.fields() {
-                decode_and_consume_type(field.ty().id(), types, input)?;
-            }
-            Ok(())
-        }
-        TypeDef::Sequence(seq) => {
-            let len = <Compact<u32>>::decode(input)?;
-            for _ in 0..len.0 {
-                decode_and_consume_type(seq.type_param().id(), types, input)?;
-            }
-            Ok(())
-        }
-        TypeDef::Array(arr) => {
-            for _ in 0..arr.len() {
-                decode_and_consume_type(arr.type_param().id(), types, input)?;
-            }
-            Ok(())
-        }
-        TypeDef::Tuple(tuple) => {
-            for field in tuple.fields() {
-                decode_and_consume_type(field.id(), types, input)?;
-            }
-            Ok(())
-        }
-        TypeDef::Primitive(primitive) => {
-            match primitive {
-                TypeDefPrimitive::Bool => consume_type::<bool>(input),
-                TypeDefPrimitive::Char => {
-                    Err(
-                        EventsDecodingError::UnsupportedPrimitive(TypeDefPrimitive::Char)
-                            .into(),
-                    )
-                }
-                TypeDefPrimitive::Str => consume_type::<String>(input),
-                TypeDefPrimitive::U8 => consume_type::<u8>(input),
-                TypeDefPrimitive::U16 => consume_type::<u16>(input),
-                TypeDefPrimitive::U32 => consume_type::<u32>(input),
-                TypeDefPrimitive::U64 => consume_type::<u64>(input),
-                TypeDefPrimitive::U128 => consume_type::<u128>(input),
-                TypeDefPrimitive::U256 => {
-                    Err(
-                        EventsDecodingError::UnsupportedPrimitive(TypeDefPrimitive::U256)
-                            .into(),
-                    )
-                }
-                TypeDefPrimitive::I8 => consume_type::<i8>(input),
-                TypeDefPrimitive::I16 => consume_type::<i16>(input),
-                TypeDefPrimitive::I32 => consume_type::<i32>(input),
-                TypeDefPrimitive::I64 => consume_type::<i64>(input),
-                TypeDefPrimitive::I128 => consume_type::<i128>(input),
-                TypeDefPrimitive::I256 => {
-                    Err(
-                        EventsDecodingError::UnsupportedPrimitive(TypeDefPrimitive::I256)
-                            .into(),
-                    )
-                }
-            }
-        }
-        TypeDef::Compact(compact) => {
-            let inner = types
-                .resolve(compact.type_param().id())
-                .ok_or(MetadataError::TypeNotFound(type_id))?;
-            let mut decode_compact_primitive = |primitive: &TypeDefPrimitive| {
-                match primitive {
-                    TypeDefPrimitive::U8 => consume_type::<Compact<u8>>(input),
-                    TypeDefPrimitive::U16 => consume_type::<Compact<u16>>(input),
-                    TypeDefPrimitive::U32 => consume_type::<Compact<u32>>(input),
-                    TypeDefPrimitive::U64 => consume_type::<Compact<u64>>(input),
-                    TypeDefPrimitive::U128 => consume_type::<Compact<u128>>(input),
-                    prim => {
-                        Err(EventsDecodingError::InvalidCompactPrimitive(prim.clone())
-                            .into())
-                    }
-                }
-            };
-            match inner.type_def() {
-                TypeDef::Primitive(primitive) => decode_compact_primitive(primitive),
-                TypeDef::Composite(composite) => {
-                    match composite.fields() {
-                        [field] => {
-                            let field_ty =
-                                types.resolve(field.ty().id()).ok_or_else(|| {
-                                    MetadataError::TypeNotFound(field.ty().id())
-                                })?;
-                            if let TypeDef::Primitive(primitive) = field_ty.type_def() {
-                                decode_compact_primitive(primitive)
-                            } else {
-                                Err(EventsDecodingError::InvalidCompactType(
-                                    "Composite type must have a single primitive field"
-                                        .into(),
-                                )
-                                .into())
-                            }
-                        }
-                        _ => {
-                            Err(EventsDecodingError::InvalidCompactType(
-                                "Composite type must have a single field".into(),
-                            )
-                            .into())
-                        }
-                    }
-                }
-                _ => {
-                    Err(EventsDecodingError::InvalidCompactType(
-                        "Compact type must be a primitive or a composite type".into(),
-                    )
-                    .into())
-                }
-            }
-        }
-        TypeDef::BitSequence(bitseq) => {
-            let bit_store_def = types
-                .resolve(bitseq.bit_store_type().id())
-                .ok_or(MetadataError::TypeNotFound(type_id))?
-                .type_def();
-
-            // We just need to consume the correct number of bytes. Roughly, we encode this
-            // as a Compact<u32> length, and then a slice of T of that length, where T is the
-            // bit store type. So, we ignore the bit order and only care that the bit store type
-            // used lines up in terms of the number of bytes it will take to encode/decode it.
-            match bit_store_def {
-                TypeDef::Primitive(TypeDefPrimitive::U8) => {
-                    consume_type::<BitVec<Lsb0, u8>>(input)
-                }
-                TypeDef::Primitive(TypeDefPrimitive::U16) => {
-                    consume_type::<BitVec<Lsb0, u16>>(input)
-                }
-                TypeDef::Primitive(TypeDefPrimitive::U32) => {
-                    consume_type::<BitVec<Lsb0, u32>>(input)
-                }
-                TypeDef::Primitive(TypeDefPrimitive::U64) => {
-                    consume_type::<BitVec<Lsb0, u64>>(input)
-                }
-                store => {
-                    return Err(EventsDecodingError::InvalidBitSequenceType(format!(
-                        "{:?}",
-                        store
-                    ))
-                    .into())
-                }
-            }
-        }
-    }
-}
-
-/// The possible errors that we can run into attempting to decode events.
-#[derive(Debug, thiserror::Error)]
-pub enum EventsDecodingError {
-    /// Unsupported primitive type
-    #[error("Unsupported primitive type {0:?}")]
-    UnsupportedPrimitive(TypeDefPrimitive),
-    /// Invalid compact type, must be an unsigned int.
-    #[error("Invalid compact primitive {0:?}")]
-    InvalidCompactPrimitive(TypeDefPrimitive),
-    /// Invalid compact type; error details in string.
-    #[error("Invalid compact composite type {0}")]
-    InvalidCompactType(String),
-    /// Invalid bit sequence type; bit store type or bit order type used aren't supported.
-    #[error("Invalid bit sequence type; bit store type {0} is not supported")]
-    InvalidBitSequenceType(String),
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::{
-        error::GenericError::{
-            Codec,
-            EventsDecoding,
-            Other,
-        },
-        events::EventsDecodingError::UnsupportedPrimitive,
         Config,
         DefaultConfig,
         Phase,
     };
-    use assert_matches::assert_matches;
     use codec::Encode;
     use frame_metadata::{
         v14::{
@@ -711,8 +364,6 @@ mod tests {
         TypeInfo,
     };
     use std::convert::TryFrom;
-
-    type TypeId = scale_info::interner::UntrackedSymbol<std::any::TypeId>;
 
     /// An "outer" events enum containing exactly one event.
     #[derive(Encode, Decode, TypeInfo, Clone, Debug, PartialEq)]
@@ -737,17 +388,6 @@ mod tests {
             event: AllEvents::E(event),
             topics: vec![],
         }
-    }
-
-    /// Build a type registry that knows about the single type provided.
-    fn singleton_type_registry<T: scale_info::TypeInfo + 'static>(
-    ) -> (TypeId, PortableRegistry) {
-        let m = scale_info::MetaType::new::<T>();
-        let mut types = scale_info::Registry::new();
-        let id = types.register_type(&m);
-        let portable_registry: PortableRegistry = types.into();
-
-        (id, portable_registry)
     }
 
     /// Build fake metadata consisting of a single pallet that knows
@@ -807,18 +447,7 @@ mod tests {
         }
     }
 
-    fn decode_and_consume_type_consumes_all_bytes<
-        T: codec::Encode + scale_info::TypeInfo + 'static,
-    >(
-        val: T,
-    ) {
-        let (type_id, registry) = singleton_type_registry::<T>();
-        let bytes = val.encode();
-        let cursor = &mut &*bytes;
 
-        decode_and_consume_type(type_id.id(), &registry, cursor).unwrap();
-        assert_eq!(cursor.len(), 0);
-    }
 
     #[test]
     fn statically_decode_single_event() {
@@ -1293,219 +922,5 @@ mod tests {
                 data: expected_event_data.into()
             }]
         );
-    }
-
-    #[test]
-    fn decode_bitvec() {
-        use bitvec::order::Msb0;
-
-        decode_and_consume_type_consumes_all_bytes(
-            bitvec::bitvec![Lsb0, u8; 0, 1, 1, 0, 1],
-        );
-        decode_and_consume_type_consumes_all_bytes(
-            bitvec::bitvec![Msb0, u8; 0, 1, 1, 0, 1, 0, 1, 0, 0],
-        );
-
-        decode_and_consume_type_consumes_all_bytes(
-            bitvec::bitvec![Lsb0, u16; 0, 1, 1, 0, 1],
-        );
-        decode_and_consume_type_consumes_all_bytes(
-            bitvec::bitvec![Msb0, u16; 0, 1, 1, 0, 1, 0, 1, 0, 0],
-        );
-
-        decode_and_consume_type_consumes_all_bytes(
-            bitvec::bitvec![Lsb0, u32; 0, 1, 1, 0, 1],
-        );
-        decode_and_consume_type_consumes_all_bytes(
-            bitvec::bitvec![Msb0, u32; 0, 1, 1, 0, 1, 0, 1, 0, 0],
-        );
-
-        decode_and_consume_type_consumes_all_bytes(
-            bitvec::bitvec![Lsb0, u64; 0, 1, 1, 0, 1],
-        );
-        decode_and_consume_type_consumes_all_bytes(
-            bitvec::bitvec![Msb0, u64; 0, 1, 1, 0, 1, 0, 1, 0, 0],
-        );
-    }
-
-    #[test]
-    fn decode_primitive() {
-        decode_and_consume_type_consumes_all_bytes(false);
-        decode_and_consume_type_consumes_all_bytes(true);
-
-        let dummy_data = vec![0u8];
-        let dummy_cursor = &mut &*dummy_data;
-        let (id, reg) = singleton_type_registry::<char>();
-        let res = decode_and_consume_type(id.id(), &reg, dummy_cursor);
-        assert_matches!(
-            res,
-            Err(EventsDecoding(UnsupportedPrimitive(TypeDefPrimitive::Char)))
-        );
-
-        decode_and_consume_type_consumes_all_bytes("str".to_string());
-
-        decode_and_consume_type_consumes_all_bytes(1u8);
-        decode_and_consume_type_consumes_all_bytes(1i8);
-
-        decode_and_consume_type_consumes_all_bytes(1u16);
-        decode_and_consume_type_consumes_all_bytes(1i16);
-
-        decode_and_consume_type_consumes_all_bytes(1u32);
-        decode_and_consume_type_consumes_all_bytes(1i32);
-
-        decode_and_consume_type_consumes_all_bytes(1u64);
-        decode_and_consume_type_consumes_all_bytes(1i64);
-
-        decode_and_consume_type_consumes_all_bytes(1u128);
-        decode_and_consume_type_consumes_all_bytes(1i128);
-    }
-
-    #[test]
-    fn decode_tuple() {
-        decode_and_consume_type_consumes_all_bytes(());
-
-        decode_and_consume_type_consumes_all_bytes((true,));
-
-        decode_and_consume_type_consumes_all_bytes((true, "str"));
-
-        // Incomplete bytes for decoding
-        let dummy_data = false.encode();
-        let dummy_cursor = &mut &*dummy_data;
-        let (id, reg) = singleton_type_registry::<(bool, &'static str)>();
-        let res = decode_and_consume_type(id.id(), &reg, dummy_cursor);
-        assert_matches!(res, Err(Codec(_)));
-
-        // Incomplete bytes for decoding, with invalid char type
-        let dummy_data = (false, "str", 0u8).encode();
-        let dummy_cursor = &mut &*dummy_data;
-        let (id, reg) = singleton_type_registry::<(bool, &'static str, char)>();
-        let res = decode_and_consume_type(id.id(), &reg, dummy_cursor);
-        assert_matches!(
-            res,
-            Err(EventsDecoding(UnsupportedPrimitive(TypeDefPrimitive::Char)))
-        );
-        // The last byte (0x0 u8) should not be consumed
-        assert_eq!(dummy_cursor.len(), 1);
-    }
-
-    #[test]
-    fn decode_array_and_seq() {
-        decode_and_consume_type_consumes_all_bytes([0]);
-        decode_and_consume_type_consumes_all_bytes([1, 2, 3, 4, 5]);
-        decode_and_consume_type_consumes_all_bytes([0; 500]);
-        decode_and_consume_type_consumes_all_bytes(["str", "abc", "cde"]);
-
-        decode_and_consume_type_consumes_all_bytes(vec![0]);
-        decode_and_consume_type_consumes_all_bytes(vec![1, 2, 3, 4, 5]);
-        decode_and_consume_type_consumes_all_bytes(vec!["str", "abc", "cde"]);
-    }
-
-    #[test]
-    fn decode_variant() {
-        #[derive(Clone, Encode, TypeInfo)]
-        enum EnumVar {
-            A,
-            B((&'static str, u8)),
-            C { named: i16 },
-        }
-        const INVALID_TYPE_ID: u32 = 1024;
-
-        decode_and_consume_type_consumes_all_bytes(EnumVar::A);
-        decode_and_consume_type_consumes_all_bytes(EnumVar::B(("str", 1)));
-        decode_and_consume_type_consumes_all_bytes(EnumVar::C { named: 1 });
-
-        // Invalid variant index
-        let dummy_data = 3u8.encode();
-        let dummy_cursor = &mut &*dummy_data;
-        let (id, reg) = singleton_type_registry::<EnumVar>();
-        let res = decode_and_consume_type(id.id(), &reg, dummy_cursor);
-        assert_matches!(res, Err(Other(_)));
-
-        // Valid index, incomplete data
-        let dummy_data = 2u8.encode();
-        let dummy_cursor = &mut &*dummy_data;
-        let res = decode_and_consume_type(id.id(), &reg, dummy_cursor);
-        assert_matches!(res, Err(Codec(_)));
-
-        let res = decode_and_consume_type(INVALID_TYPE_ID, &reg, dummy_cursor);
-        assert_matches!(res, Err(crate::error::GenericError::Metadata(_)));
-    }
-
-    #[test]
-    fn decode_composite() {
-        #[derive(Clone, Encode, TypeInfo)]
-        struct Composite {}
-        decode_and_consume_type_consumes_all_bytes(Composite {});
-
-        #[derive(Clone, Encode, TypeInfo)]
-        struct CompositeV2 {
-            id: u32,
-            name: String,
-        }
-        decode_and_consume_type_consumes_all_bytes(CompositeV2 {
-            id: 10,
-            name: "str".to_string(),
-        });
-
-        #[derive(Clone, Encode, TypeInfo)]
-        struct CompositeV3<T> {
-            id: u32,
-            extra: T,
-        }
-        decode_and_consume_type_consumes_all_bytes(CompositeV3 {
-            id: 10,
-            extra: vec![0, 1, 2],
-        });
-        decode_and_consume_type_consumes_all_bytes(CompositeV3 {
-            id: 10,
-            extra: bitvec::bitvec![Lsb0, u8; 0, 1, 1, 0, 1],
-        });
-        decode_and_consume_type_consumes_all_bytes(CompositeV3 {
-            id: 10,
-            extra: ("str", 1),
-        });
-        decode_and_consume_type_consumes_all_bytes(CompositeV3 {
-            id: 10,
-            extra: CompositeV2 {
-                id: 2,
-                name: "str".to_string(),
-            },
-        });
-
-        #[derive(Clone, Encode, TypeInfo)]
-        struct CompositeV4(u32, bool);
-        decode_and_consume_type_consumes_all_bytes(CompositeV4(1, true));
-
-        #[derive(Clone, Encode, TypeInfo)]
-        struct CompositeV5(u32);
-        decode_and_consume_type_consumes_all_bytes(CompositeV5(1));
-    }
-
-    #[test]
-    fn decode_compact() {
-        #[derive(Clone, Encode, TypeInfo)]
-        enum Compact {
-            A(#[codec(compact)] u32),
-        }
-        decode_and_consume_type_consumes_all_bytes(Compact::A(1));
-
-        #[derive(Clone, Encode, TypeInfo)]
-        struct CompactV2(#[codec(compact)] u32);
-        decode_and_consume_type_consumes_all_bytes(CompactV2(1));
-
-        #[derive(Clone, Encode, TypeInfo)]
-        struct CompactV3 {
-            #[codec(compact)]
-            val: u32,
-        }
-        decode_and_consume_type_consumes_all_bytes(CompactV3 { val: 1 });
-
-        #[derive(Clone, Encode, TypeInfo)]
-        struct CompactV4<T> {
-            #[codec(compact)]
-            val: T,
-        }
-        decode_and_consume_type_consumes_all_bytes(CompactV4 { val: 0u8 });
-        decode_and_consume_type_consumes_all_bytes(CompactV4 { val: 1u16 });
     }
 }
